🛠 fixes cross-ref bug #59
